pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "perf_agent"
version = "0.1.0"
description = "Central-state-machine based performance analysis scaffold"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=2.30,<3.0",
  "pydantic>=2.6,<3.0",
  "python-dotenv>=1.0,<2.0",
  "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[project.scripts]
perf-agent = "perf_agent.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
